refactor(MoveDto) : Vo 역할을 Dto 역할로 정정
