+++ src/main/java/chess/domain/dto/MoveDto.java
@@ -1,6 +1,6 @@
-package chess.domain.vo;
+package chess.domain.dto;
 
-public class MoveVo {
+public class MoveDto {
     private String source;
     private String target;
     private String gameId;
